container.cのruntest()内の警告を修正

+++ container.c
@@ -26,8 +26,7 @@ void runtest() {
     expect(__LINE__, 0, vec->len);
 
     for (int i = 0; i < 100; i++)
-        // TODO: -Wint-to-pointer-cast
-        vec_push(vec, (void *)i);
+        vec_push(vec, (void *)(long)i);
 
     expect(__LINE__, 100, vec->len);
     expect(__LINE__, 0, (long)vec->data[0]);
